v1.0.7
Converted the locations from a field group in the Trial to its own post type.

Added NCTID, Status and Languages taxonomy for new locations post type.

Reworked the import to start with the trial, then import it's locations.

Added new fields for locations, and a new meta box to fetch and save the latitude and longitude for the location.

Updated composer dependencies.